chore: remove service name details from events and function handler naming scheme
