+++ apps/course-builder-web/src/inngest/video-processing/transcript-ready.ts
@@ -2,8 +2,8 @@ import { db } from '@/db'
 import { contentResource } from '@/db/schema'
 import { env } from '@/env.mjs'
 import { inngest } from '@/inngest/inngest.server'
-import { DEEPGRAM_WEBHOOK_EVENT } from '@/inngest/video-processing/events/video-deepgram-webhook'
 import { VIDEO_SRT_READY_EVENT } from '@/inngest/video-processing/events/video-srt-ready-to-asset'
+import { VIDEO_TRANSCRIPT_READY_EVENT } from '@/inngest/video-processing/events/video-transcript-ready'
 import {
   srtFromTranscriptResult,
   transcriptAsParagraphsWithTimestamps,
@@ -12,9 +12,9 @@ import {
 import { sql } from 'drizzle-orm'
 
 export const deepgramTranscriptReady = inngest.createFunction(
-  { id: `deepgram-transcript-ready-event`, name: 'Deepgram Transcript Ready' },
+  { id: `transcript-ready-event`, name: 'Transcript Ready' },
   {
-    event: DEEPGRAM_WEBHOOK_EVENT,
+    event: VIDEO_TRANSCRIPT_READY_EVENT,
   },
   async ({ event, step }) => {
     const videoResourceId = event.data.videoResourceId
@@ -33,7 +33,7 @@ export const deepgramTranscriptReady = inngest.createFunction(
           ${contentResource.fields} = JSON_SET(
             ${contentResource.fields}, '$.transcript', ${transcript}),
           ${contentResource.fields} = JSON_SET(
-            ${contentResource.fields}, '$.srt', ${srt}),
+            ${contentResource.fields},
           ${contentResource.fields} = JSON_SET(
             ${contentResource.fields}, '$.wordLevelSrt', ${wordLevelSrt}),
           ${contentResource.fields} = JSON_SET(
